# backend/scripts/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Campus Operations AI API",
                 "description":  "API for Campus Operations Intelligence Platform",
                 "version":  "0.0.1"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
    "paths":  {
...
                  "/api/schedules/{id}":  {
                                              "get":  {
                                                          "tags":  [
                                                                       "schedule-controller"
                                                                   ],
                                                          "operationId":  "getScheduleById",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "id",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  "@{schema=}"
                                                                                                    }
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "bearerAuth":  [

                                                                                              ]
                                                                           }
                                                                       ]
                                                      },
                                              "put":  {
                                                          "tags":  [
                                                                       "schedule-controller"
                                                                   ],
                                                          "operationId":  "updateSchedule",
                                                          "parameters":  [
                                                                             {
                                                                                 "name":  "id",
                                                                                 "in":  "path",
                                                                                 "required":  true,
                                                                                 "schema":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            }
                                                                             }
                                                                         ],
                                                          "requestBody":  {
                                                                              "content":  {
                                                                                              "application/json":  {
                                                                                                                       "schema":  "@{$ref=#/components/schemas/UpdateScheduleRequest}"
                                                                                                                   }
                                                                                          },
                                                                              "required":  true
                                                                          },
                                                          "responses":  {
                                                                            "200":  {
                                                                                        "description":  "OK",
                                                                                        "content":  {
                                                                                                        "*/*":  "@{schema=}"
                                                                                                    }
                                                                                    }
                                                                        },
                                                          "security":  [
                                                                           {
                                                                               "bearerAuth":  [

                                                                                              ]
                                                                           }
                                                                       ]
                                                      },
                                              "delete":  {
                                                             "tags":  [
                                                                          "schedule-controller"
                                                                      ],
                                                             "operationId":  "deleteSchedule",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  "@{schema=}"
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                          },
...
                  "/api/schedules":  {
                                         "get":  {
                                                     "tags":  [
                                                                  "schedule-controller"
                                                              ],
                                                     "operationId":  "getAllSchedules",
                                                     "responses":  {
                                                                       "200":  {
                                                                                   "description":  "OK",
                                                                                   "content":  {
                                                                                                   "*/*":  "@{schema=}"
                                                                                               }
                                                                               }
                                                                   },
                                                     "security":  [
                                                                      {
                                                                          "bearerAuth":  [

                                                                                         ]
                                                                      }
                                                                  ]
                                                 },
                                         "post":  {
                                                      "tags":  [
                                                                   "schedule-controller"
                                                               ],
                                                      "operationId":  "createSchedule",
                                                      "requestBody":  {
                                                                          "content":  {
                                                                                          "application/json":  {
                                                                                                                   "schema":  "@{$ref=#/components/schemas/CreateScheduleRequest}"
                                                                                                               }
                                                                                      },
                                                                          "required":  true
                                                                      },
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK",
                                                                                    "content":  {
                                                                                                    "*/*":  "@{schema=}"
                                                                                                }
                                                                                }
                                                                    },
                                                      "security":  [
                                                                       {
                                                                           "bearerAuth":  [

                                                                                          ]
                                                                       }
                                                                   ]
                                                  }
                                     },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "UpdateScheduleRequest":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "courseId":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int64"
                                                                                                     },
                                                                                        "roomId":  {
                                                                                                       "type":  "integer",
                                                                                                       "format":  "int64"
                                                                                                   },
                                                                                        "day":  {
                                                                                                    "type":  "string",
                                                                                                    "enum":  [
                                                                                                                 "MONDAY",
                                                                                                                 "TUESDAY",
                                                                                                                 "WEDNESDAY",
                                                                                                                 "THURSDAY",
                                                                                                                 "FRIDAY",
                                                                                                                 "SATURDAY",
                                                                                                                 "SUNDAY"
                                                                                                             ]
                                                                                                },
                                                                                        "startTime":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                        "endTime":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                    },
                                                                     "required":  [
                                                                                      "courseId",
                                                                                      "day",
                                                                                      "endTime",
                                                                                      "roomId",
                                                                                      "startTime"
                                                                                  ]
                                                                 },
...
                                       "CreateScheduleRequest":  {
                                                                     "type":  "object",
                                                                     "properties":  {
                                                                                        "courseId":  {
                                                                                                         "type":  "integer",
                                                                                                         "format":  "int64"
                                                                                                     },
                                                                                        "roomId":  {
                                                                                                       "type":  "integer",
                                                                                                       "format":  "int64"
                                                                                                   },
                                                                                        "day":  {
                                                                                                    "type":  "string",
                                                                                                    "enum":  [
                                                                                                                 "MONDAY",
                                                                                                                 "TUESDAY",
                                                                                                                 "WEDNESDAY",
                                                                                                                 "THURSDAY",
                                                                                                                 "FRIDAY",
                                                                                                                 "SATURDAY",
                                                                                                                 "SUNDAY"
                                                                                                             ]
                                                                                                },
                                                                                        "startTime":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                        "endTime":  {
                                                                                                        "type":  "string"
                                                                                                    }
                                                                                    },
                                                                     "required":  [
                                                                                      "courseId",
                                                                                      "day",
                                                                                      "endTime",
                                                                                      "roomId",
                                                                                      "startTime"
                                                                                  ]
                                                                 },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}